[280453] [performance] DefaultFileServiceCodePageConverter is wasteful with main memory

+++ rse/plugins/org.eclipse.rse.services/src/org/eclipse/rse/services/files/DefaultFileServiceCodePageConverter.java
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2007, 2011 IBM Corporation. All rights reserved.
+ * Copyright (c) 2007, 2012 IBM Corporation. All rights reserved.
  * This program and the accompanying materials are made available under the terms
  * of the Eclipse Public License v1.0 which accompanies this distribution, and is
  * available at http://www.eclipse.org/legal/epl-v10.html
@@ -93,7 +93,10 @@ public class DefaultFileServiceCodePageConverter implements
 					}
 					inputStream.close();	
 					outStream.close();
-					altFile.renameTo(file);
+					if(!altFile.renameTo(file) && file.exists()) {
+						file.delete();
+						altFile.renameTo(file);
+					}		
 				}
 
 			}
